:hammer: further optimization. all jinja2 related code should go in jinja2 folder
